fix(chat): move slash catalog types out of tui
