+++ src/chat/tui.rs
@@ -158,17 +158,7 @@ pub struct SlashMenuEntry {
     pub append_space: bool,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SlashProviderModelCatalog {
-    pub provider: String,
-    pub models: Vec<SlashModelCandidate>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SlashModelCandidate {
-    pub name: String,
-    pub description: String,
-}
+pub use crate::chat::slash_types::{SlashModelCandidate, SlashProviderModelCatalog};
 
 pub(crate) struct SlashMenuSources<'a> {
     pub live_sessions: &'a [crate::chat::sessions::SwitcherEntry],
